\name{stmBrowser-package}
\alias{stmBrowser-package}
\docType{package}
\title{
  STM Browser
}
\description{
  This package creates a browser visualization of the topics and
  covariate relationships computed from the Structural Topic Model. The
  most important and only function in the package is
  \code{\link{stmBrowser}}, which creates and opens a browser that the
  user can use to explore the output from the STM.
}
\author{
Author: Michael K. Freeman, Jason Chuang, Margaret E. Roberts, Brandon
M. Stewart, and Dustin Tingley.

Maintainer: Margaret E. Roberts <meroberts@ucsd.edu>
}
\references{
  Roberts, M., Stewart, B., Tingley, D., and Airoldi, E. (2013)
  "The
  structural topic model and applied social science." In Advances in
  Neural Information Processing Systems Workshop on Topic Models:
  Computation, Application, and Evaluation. http://goo.gl/uHkXAQ

  Roberts, M., Stewart, B., Tingley, D., Lucas, C., Leder-Luis, J.,
  Gadarian, S., Albertson, B., Albertson, B. and Rand,
  D. (Forthcoming). "Structural topic models for open ended survey
  responses." American Journal of Political Science http://goo.gl/0x0tHJ
  }                        
\keyword{ package }
\seealso{
  \code{\link{stmBrowser}}
}
