% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stmBrowser_widget.R
\name{stmBrowser_widget}
\alias{stmBrowser_widget}
\title{Create an htmlwidget view of STM}
\usage{
stmBrowser_widget(mod, data, covariates, text, id = NULL, n = 1000,
  labeltype = "prob", width = NULL, height = NULL)
}
\arguments{
\item{mod}{stm model}

\item{data}{data for stm model}

\item{covariates}{vector of covariates you want to visualize}

\item{text}{name of covariate where text is held}
}
\value{
An object of class \code{htmlwidget} that will
  intelligently print itself into HTML in a variety of contexts
  including the R console, within R Markdown documents,
  and within Shiny output bindings.
}
\description{
Create an htmlwidget view of STM
}
\examples{
\dontrun{
library(stm)
data(poliblog5k)
#Create date
dec312007 <- as.numeric(as.Date("2007-12-31"))
poliblog5k.meta$date <- as.Date(dec312007+poliblog5k.meta$day,
                               origin="1970-01-01")
out <- prepDocuments(poliblog5k.docs, poliblog5k.voc, poliblog5k.meta)
stm.out <- stm(out$documents, out$vocab, K=50,
              prevalence=~rating + date,
              data=out$meta, init.type="Spectral",
              max.em.its=10) #generally run models longer than this.
library(stmBrowser)
stmBrowser_widget(stm.out, data=out$meta, c("rating", "date"),
          text="text")
}
}

