% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stmBrowser_widget.R
\name{stmBrowser_widget}
\alias{stmBrowser_widget}
\title{Function for outputing STM dataset into json}
\usage{
stmBrowser_widget(mod, data, covariates, text, id = NULL, n = 1000,
  labeltype = "prob")
}
\arguments{
\item{mod}{stm model}

\item{data}{data for stm model}

\item{covariates}{vector of covariates you want to visualize}

\item{text}{name of covariate where text is held}
}
\value{
An object of class \code{htmlwidget} that will
  intelligently print itself into HTML in a variety of contexts
  including the R console, within R Markdown documents,
  and within Shiny output bindings.
}
\description{
Function for outputing STM dataset into json
}

